mechanistic/requirements.txt
transformers>=4.41.0
torch>=2.2.0
torchvision>=0.17.0
tqdm
pandas
numpy
scikit-learn
matplotlib
opencv-python
pyyaml
sentencepiece
